requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
boto3==1.26.109
botocore==1.29.109
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cssselect==1.2.0
Django==3.2.5
numpy==1.21.6
pandas==1.3.5
exceptiongroup==1.1.1
fake-useragent==1.1.1
filelock==3.11.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud==0.34.0
google-cloud-vision==3.3.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.0.6
jmespath==1.0.1
loguru==0.6.0
lxml==4.9.2
pymysql==1.0.3
outcome==1.2.0
packaging==23.0
parsel==1.7.0
Protego==0.2.1
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rsa==4.9
s3transfer==0.6.0
Scrapy==2.5.1
scrapy-djangoitem==1.1.1
scrapy-selenium==0.0.7
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
w3lib==2.1.1
wsproto==1.2.0
zipp==3.15.0
zope.interface==6.0
gunicorn==20.1.0
scikit-learn==1.0.2
pyopenssl==22.0.0
cryptography==37.0.0
xgboost
dill